src/requirements.txt
asyncio==3.4.3
asyncpg==0.23.0
aiodocker==0.21.0
fastapi==0.65.1
uvicorn==0.14.0
python-dateutil==2.8.1
httpx==0.25.0
pytest==3.10.1
requests==2.26.0